requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.6.15
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.29.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-import-export==3.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mypy-extensions==0.4.3
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.2
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.12
waitress==2.1.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0